Graphviz fixes
I was able to change the nodes to a point and have all the edges have a
color. Although, I cannot figure out how to color of the points. The
points are yet still too big to visualize the graph the same as they
did in the graphviz tutorial. The color *should* be easily visible
though for partitioning purposes.